+++ app/server/server_plot.R
@@ -2,27 +2,24 @@
 
 observeEvent({
   
-  c(input$tabs) # , input$input_files
+  c(input$tabs, input$radio_period)
 },
 {
   if(is.null(db)) {
     return(NULL)
   } else {
+    # Compute number of days
+    x <- db %>% mutate(d_days = as.integer(difftime(max(date), date, units = "days")))
+    if(input$radio_period == "Week") {
+      x <- x %>% filter(d_days <= 7)
+    } else if(input$radio_period == "Month") {
+      x <- x %>% filter(d_days <= 31)
+    } else {
+      x <- x %>% filter(d_days <= 365)
+    }
     
-    p <- ggplot(db %>% mutate(instr = factor(instr)),
-                aes(x = date, y = ct, colour = target)) +
-      geom_point(aes(shape = instr), alpha = 0.5) + #
-      geom_line() +
-      labs(x = "Date",
-           colour = "Target name",
-           shape = "Instrument") +
-      scale_color_manual(values=colors) +
-      facet_wrap(~target, ncol = 5, scales = "fixed") +
-      theme_bw() +
-      theme(legend.position="none")
-    
-    output$plot_trend <- renderPlotly({
-      ggplotly(p)
-    })
+    p <- plotTrend(x)
+
+    output$plot_trend <- renderPlotly(ggplotly(p, height=1600))
   }
 })