+++ CHANGELOG.md
@@ -6,6 +6,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.3] - 2021-08-06
+### Added
+- README documentation
+- Radio button to select time period
+
+### Changed
+- Plot output
+- Main text as the db_path is not necessary anymore
+
+### Removed
+- Both the targets and db_path as global parameters
+
+## [0.0.2] - 2021-08-05
+### Added
+- README documentation
+- Trend line
+
+### Changed
+- Plot output
+
+### Removed
+- Targets as global input
+
 ## [0.0.2] - 2021-08-05
 
 ### Added